tools: using command args for build scripts
